+++ analysis/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 - Add new '--base-file' flag to unifiedparser and rawtextparser [#4270](https://github.com/MaibornWolff/codecharta/pull/4270)
 - UnifiedParser now automatically uses `.gitignore` files for file exclusion [#4254](https://github.com/MaibornWolff/codecharta/issues/4254)
 - RawTextParser now automatically uses `.gitignore` files for file exclusion [#4273](https://github.com/MaibornWolff/codecharta/issues/4273)
+- Add swift support to unifiedparser [#4335](https://github.com/MaibornWolff/codecharta/issues/4335)
 - GitLogParser is now able to find commits that contain the word hotfixes
 - Add two code smell metrics to unifiedparser [#4315](https://github.com/MaibornWolff/codecharta/issues/4315)
   - long function